feat(kernel-adoption): E4 — exact Money/FX boundary adapter
One adapter between kernel Money/Currency/ExchangeRate values and ERP's
Decimal + currency_code contracts (app/services/finance/money_boundary.py):
exact conversion and byte-exact connector serialization, the slice's single
centralized rounding decision (half-up minor units), fail-closed validators
(float/bool, missing or invalid currency, currency mismatch, excess
minor-unit precision), the WHT settlement identity with the pre-existing
one-minor-unit tolerance, and FX snapshots built ONLY from persisted
ERP-owned core_fx.exchange_rate observations (pair, rate type, effective
time, source, row identity) — never a live lookup in a posting path.

Used first at the Sub-facing boundary: InvoiceRecord / CreditNoteRecord /
PaymentRecord gain fail-closed boundary_money() (headers, WHT evidence,
allocations), enforced per-row inside the sync savepoints; the Sub/CRM
payables command CRMPurchaseInvoicePayload validates header totals and line
amounts through the adapter at parse time. The _CENTS/ROUND_HALF_UP
scattering and the flat 0.01 WHT tolerance are replaced by currency-aware
adapter equivalents (behavior-identical for 2-minor-unit currencies).

ERP internals unchanged (boundary 6): Numeric(20,6) posting/functional
amounts, Numeric(20,10) FX rates, tax ratios, quantities/unit prices,
posted-line snapshots, account mappings. Golden money pins stay
byte-identical green. Float eliminations (ledger finding 7): fixed-assets
maintenance work order estimated_cost/actual_cost now Mapped[Decimal]
(columns were already Numeric(20,6); no schema change) and the float()
writes in the maintenance service removed.

The E2 app-unchanged canary evolves into
test_app_import_loads_only_pure_contract_kernel_modules: app.main may load
only the consume-pure kernel import closure; kernel db/messaging/session
surfaces still fail loudly.

+++ app/models/fixed_assets/maintenance_work_order.py
@@ -5,6 +5,7 @@ Fixed Asset Maintenance Work Order model.
 import enum
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import (
     DateTime,
@@ -110,13 +111,16 @@ class MaintenanceWorkOrder(Base):
     )
     completion_notes: Mapped[str | None] = mapped_column(Text, nullable=True)
 
-    estimated_cost: Mapped[float] = mapped_column(
+    # Money columns are exact decimals (kernel-adoption E4 float elimination);
+    # SQLAlchemy already returned Decimal for Numeric — the float annotations
+    # were a lie. No schema change.
+    estimated_cost: Mapped[Decimal] = mapped_column(
         Numeric(20, 6),
         nullable=False,
         default=0,
         server_default="0",
     )
-    actual_cost: Mapped[float] = mapped_column(
+    actual_cost: Mapped[Decimal] = mapped_column(
         Numeric(20, 6),
         nullable=False,
         default=0,
